docs: moved wallet based doc tests, from fuel-guage to doc snippets
