+++ .changeset/blue-hotels-heal.md
@@ -1,0 +1,4 @@
+---
+---
+
+Moved wallet doc snippets from `fuel-gauge` to `docs-snippets` and updated documentation.
