+++ scripts/developer.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+pnpm lint:package-jsons
+pnpm forc:check
+pnpm lint
+pnpm test:validate
+# pnpm test
